fix bug caused by the rebase
